chore(google): Bump google compute dependencies (#306)
We're currently using an alpha version of the client library for
GCE from December 2017; all of the features we were using are in
the current version of the beta client library.

Given that we aren't using any alpha features, let's move to the
beta version of the library. If we want to support an alpha feature,
it's easy to move back to the alpha version.