feat(*): Fail the build unless the .template and .yml file are in sync (#310)
It's extremely common for people to update spinnaker-dependencies.template
but forget to update spinnaker-dependencies.yml (or vice versa). The
build task currently updates the .yml to match the .template, but
having a 'build' task modify the source code means that CI is testing
changes that aren't (and won't be committed).

Add a new 'checkDependencies' task that fails if the .template and
.yml are out of sync; make the check task depend on this task to
quickly detect when these get out of sync.